requirements.txt
hikari[speedups]==2.0.0.dev108
#hikari-lightbulb[crontrigger]==2.2.1
git+https://github.com/tandemdude/hikari-lightbulb#egg=hikari-lightbulb[crontrigger]

aiofiles==0.8.0
cassandra-driver==3.25.0
shortuuid==1.1.0

uvloop>=0.16; sys_platform != "win32"
aiohttp==3.8.1
toml==0.10.2

lavasnek-rs==0.1.0a5
yt-dlp==2022.4.8
